Re-structure: move search utils tasks into `...tasks.search`
I wasn't able to find nor solve the problem with the `cyclic-import` error
reported by pylint. So, I tried moving the utils functions used by search tasks
out from `readthedocs.projects.tasks.utils` into
`readthedocs.projects.tasks.search` instead.

This made `prospector` and `pylint` to pass locally without reporting any issue.
